When auth_redirect() detects a logged out user and the target page was about.php?updated, then issue a message welcoming them.
This is to prevent it from being so jolting if you are taken to
the login screen after an update.

In WordPress 3.4, the changes to wp_salt() provide for extra
security, but will cause a log-out for any installs without 8
unique keys and salts in wp-config.php (with some exceptions).
Properly re-issuing cookies, even for the logged in user, is
not easily doable via admin/includes/update-core.php, as that
file is included long after the headers are sent.

see #19599.



git-svn-id: https://core.svn.wordpress.org/trunk@20887 1a063a9b-81f0-0310-95a4-ce76da25c4cd