Stabilize how WordPress hooks into SimplePie to implement transient caching.
Since a plugin can load a previous (< 1.3) version of SimplePie before we do,
we need to be compatible with our old method of overriding SimplePie_Cache::create().

SimplePie_Cache::create() was converted to static in 1.3 (as it was called),
requiring that we create two different definitions of WP_Feed_Cache (extends
SimplePie_Cache). Instead, we can use 1.3's new object registry, and leave
the old WP_Feed_Cache to SimplePie 1.2 versions.

see #21183.



git-svn-id: https://core.svn.wordpress.org/trunk@21652 1a063a9b-81f0-0310-95a4-ce76da25c4cd